changed return value of GetNumScoreComponents() to size_t to match virt.func.declaration in ScoreProducer.h
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/trunk/moses@896 1f5c12ca-751b-0410-a591-d2e778427230

+++ src/DummyScoreProducers.cpp
@@ -9,7 +9,7 @@ DistortionScoreProducer::DistortionScoreProducer()
 	const_cast<ScoreIndexManager&>(StaticData::Instance()->GetScoreIndexManager()).AddScoreProducer(this);
 }
 
-unsigned int DistortionScoreProducer::GetNumScoreComponents() const
+size_t DistortionScoreProducer::GetNumScoreComponents() const
 {
 	return 1;
 }
@@ -37,7 +37,7 @@ WordPenaltyProducer::WordPenaltyProducer()
 	const_cast<ScoreIndexManager&>(StaticData::Instance()->GetScoreIndexManager()).AddScoreProducer(this);
 }
 
-unsigned int WordPenaltyProducer::GetNumScoreComponents() const
+size_t WordPenaltyProducer::GetNumScoreComponents() const
 {
 	return 1;
 }
@@ -48,7 +48,7 @@ const std::string WordPenaltyProducer::GetScoreProducerDescription() const
 }
 
 
-unsigned int UnknownWordPenaltyProducer::GetNumScoreComponents() const
+size_t UnknownWordPenaltyProducer::GetNumScoreComponents() const
 {
 	return 1;
 }
